requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.1
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0